+++ public/src/models/banner.ts
@@ -47,5 +47,4 @@ export interface BannerTest extends Test {
   deviceType?: DeviceType;
   campaignName?: string;
   contextTargeting: PageContextTargeting;
-  isBanditTest?: boolean;
 }
