+++ public/src/models/epic.ts
@@ -64,5 +64,4 @@ export interface EpicTest extends Test {
   controlProportionSettings?: ControlProportionSettings;
   deviceType?: DeviceType;
   campaignName?: string;
-  isBanditTest?: boolean;
 }
